Abstract buildFrame
